FIX: do not import ResolveGenericClassParameters for Python 3.6
